requirements.txt
opencv-python==4.10.0.84
mediapipe==0.10.14
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
joblib==1.4.2
streamlit==1.39.0
pillow==10.4.0
